Improve subclassing, so that duplicate code is eliminated
